% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fungai-package.R
\docType{package}
\name{fungai-package}
\alias{fungai-package}
\title{fungai: fungai is an ai mushroom assistant}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

fungai has been system prompted with information on 237 distinct mushrooms in addition to the LLMs training. given information or a photo, fungai should be able to help you identify said mushroom. this is an experimental library for testing edge focused LLMs as specialized assistants.
}
\author{
\strong{Maintainer}: frank hull \email{frankiethull@gmail.com}

}
\keyword{internal}
