% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mushroom_dictionary.R
\docType{data}
\name{mushroom_dictionary}
\alias{mushroom_dictionary}
\title{Mushroom Dictionary}
\format{
\subsection{\code{mushroom_dictionary}}{

A data frame with 1185 obs of 3 columns which contains mushroom descriptions
\describe{
\item{mushroom_name}{mushroom name}
\item{category}{5 unique information categories}
\item{information}{information on category}
...
}
}
}
\source{
\url{https://www.mushroom.world/mushrooms/namelist}
}
\usage{
mushroom_dictionary
}
\description{
a large dataset on mushrooms
}
\keyword{datasets}
