% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fungai.R
\name{fungai}
\alias{fungai}
\title{fungai}
\usage{
fungai(
  llm_model = "llama3.2-vision",
  prompt = "hello, fungai!! tell me something about morels!"
)
}
\arguments{
\item{llm_model}{a llm model}

\item{prompt}{a prompt for the model}

\item{...}{pass thru for chat}
}
\value{
fungai responses
}
\description{
the basic method to talk to fungai
}
